homeDirSafe option for startProcess() method
LDAP backed Unix hosts might not allow users to run commands if they
are not logged in to that host previously since they don't have a home
directory yet. This option, when set to true, creates a shell before
running a command and immediately closes it to trigger home directory
creation.